Move SearchConfiguration to separate file for better organization
