+++ src/MotorcycleRAG.Core/Models/Configuration.cs
@@ -74,24 +74,6 @@ public class RetryConfiguration
 }
 
 /// <summary>
-/// Search service configuration
-/// </summary>
-public class SearchConfiguration
-{
-    [Required]
-    public string IndexName { get; set; } = "motorcycle-index";
-
-    [Range(1, 1000)]
-    public int BatchSize { get; set; } = 100;
-
-    [Range(1, 100)]
-    public int MaxSearchResults { get; set; } = 50;
-
-    public bool EnableHybridSearch { get; set; } = true;
-    public bool EnableSemanticRanking { get; set; } = true;
-}
-
-/// <summary>
 /// Web search configuration
 /// </summary>
 public class WebSearchConfiguration
